Analyzer - Statistical Model Improvements
Several improvements had to be done so that the new precise model
can be implemented as a child class of the statistical model.

Parameter "sort" was introduced in the init. If the sorting is required
it must be done before converting IP addresses from string to object.

Sorting after the conversion is not possible mainly because both
IPv6 and IPv4 addresses can present in the traffic and they have
different data types.