CMake - avoid using -Werror when building yaml-cpp
We need to avoid using "-Werror" but rpmbuild tool adds additional
"-Werror=format-security" flag when building RPM package that we
would like to preserve.

Unfortunately, CMake does not support negative lookahead in regular
expressions, so we have to match a simple expression with a space
at the end. To make this work for situations where the parameter is
last, we always have to extend the string to be edited as well.