ft-generator - make TCP layer not choose direction to data packets so it can be chosen by further layers
Previously, the TCP layer was in charge of assigning direction to all packets.

This was okay, because there were no application layers, so TCP was always the last layer,
and it was also necessary back when ACK packets were inserted seperately from data packets,
but this was recently changed.

With this change TCP layer doesn't specify direction of data packets, only handshake packets,
which allows further application level protocol layers or the flow planner to select the desired
directions instead.