ft-generator - increase LFSR period from 2^N-1 to 2^N
The LFSR is unable to reach the all zeros state under normal operation,
which makes the period 2^N-1, meanwhile there are 2^N possible states.

This missing state can be significant when N is very small.

We circumvent this by inserting the all zeros state "manually" after
we exhaust all the possible states in a period.